+++ types/store.go
@@ -1,5 +1,5 @@
 package types
 
 type Record interface {
-	RecordID() int
+	RecordID() int64
 }
